% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{anno_hmdb}
\alias{anno_hmdb}
\title{Query metabolite ID in HMDB database}
\usage{
anno_hmdb(object, ncpus = 1)
}
\arguments{
\item{object}{a vector of HMDB IDs to query.}

\item{ncpus}{number of cpu}
}
\value{
A list of query results.
}
\description{
Query metabolite ID in HMDB database
}
