% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{assayData<-}
\alias{assayData<-}
\alias{assayData<-,Metabolite-method}
\title{set assayData}
\usage{
assayData(object) <- value

\S4method{assayData}{Metabolite}(object) <- value
}
\arguments{
\item{object}{A Metabolite object.}

\item{value}{The new assayData.}
}
\value{
A data.table of assayData.
}
\description{
Accessors for Metabolite object. `assayData<-` will update the assayData in the Metabolite object.
}
