% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch_norm_df}
\alias{batch_norm_df}
\title{batch normalization for a data.frame}
\usage{
batch_norm_df(object, test = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A data.frame object. The first two columns are "ID" and "Batch", the remaining columns for batch normalization}

\item{test}{test the function for the first 20 columns.}

\item{verbose}{print log information.}
}
\value{
A data.frame object after normalization.
}
\description{
Normalization data by the median value of each batch in a data.frame
}
