% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{column_missing_rate}
\alias{column_missing_rate}
\alias{column_missing_rate.default}
\alias{column_missing_rate.Metabolite}
\title{column missing rate}
\usage{
column_missing_rate(object)

\method{column_missing_rate}{default}(object)

\method{column_missing_rate}{Metabolite}(object)
}
\arguments{
\item{object}{An object, data.frame, data.table or Metabolite.}
}
\value{
Returns a vector of the missing rate for each column

A data.table of column missing rate.
}
\description{
Calculate column missing rate -- metabolite missingness.
}
\examples{
# for a Metabolite object
data(df_plasma)
v <- column_missing_rate(df_plasma)

}
