% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{correlation}
\alias{correlation}
\title{correlation of features between two Metabolite objects}
\usage{
correlation(
  object_X = NULL,
  object_Y = NULL,
  method = "pearson",
  verbose = TRUE
)
}
\arguments{
\item{object_X}{The first Metabolite object.}

\item{object_Y}{The second Metabolite object.}

\item{method}{a character string to calculate correlation coefficient. One of "pearson" (default), "kendall", or "spearman".}

\item{verbose}{print log information.}
}
\description{
Calculate the correlation of features between two Metabolite objects
}
\examples{
\dontrun{

r <- correlation(object_X, object_Y)

}

}
\seealso{
\code{\link{cor}}
}
