% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_plasma}
\alias{df_plasma}
\title{Example data.}
\format{
An object of class \code{Metabolite} of length 1.
}
\usage{
data(df_plasma)
}
\description{
A dataset containing 356 samples and 758 features.
}
\keyword{datasets}
