% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{featureData}
\alias{featureData}
\alias{featureData,Metabolite-method}
\title{get featureData}
\usage{
featureData(object)

\S4method{featureData}{Metabolite}(object)
}
\arguments{
\item{object}{A Metabolite object.}
}
\description{
Accessors for Metabolite object. Get the featureData in the Metabolite object.
}
