% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{filter_column_constant}
\alias{filter_column_constant}
\alias{calculate_column_constant}
\alias{filter_column_constant.default}
\alias{filter_column_constant.Metabolite}
\title{filter columns if values are constant}
\usage{
filter_column_constant(object, verbose)

calculate_column_constant(object, verbose = TRUE)

\method{filter_column_constant}{default}(object, verbose = TRUE)

\method{filter_column_constant}{Metabolite}(object, verbose = TRUE)
}
\arguments{
\item{object}{An object, data.frame, data.table or Metabolite.}

\item{verbose}{print log information.}
}
\description{
Remove columns if values are constant
}
\examples{
data(df_plasma)
v <- filter_column_constant(df_plasma)

}
