% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{impute}
\alias{impute}
\alias{impute.Metabolite}
\alias{impute.default}
\title{impute missing values}
\usage{
impute(object, method)

\method{impute}{Metabolite}(object, method = "half-min")

\method{impute}{default}(object, method = "half-min")
}
\arguments{
\item{object}{An object, a vector, data.frame, data.table or Metabolite.}

\item{method}{Imputation method, the default method is half the minimum value (`half-min`) of the metabolite. Currently support 'half-min', "median", "mean", "zero".}
}
\description{
impute missing values
}
\examples{


\dontrun{
d <- impute(object)
}

}
