% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{inverse_rank_transform}
\alias{inverse_rank_transform}
\title{rank-based inverse normal transformation}
\usage{
inverse_rank_transform(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector after transformation.
}
\description{
rank-based inverse normal transformation for a metabolite.
}
