% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{is_outlier}
\alias{is_outlier}
\title{is outlier}
\usage{
is_outlier(object, nSD = 5)
}
\arguments{
\item{object}{An object, a vector.}

\item{nSD}{N times of the SD as outliers.}
}
\value{
TRUE or FALSE for a vector.
}
\description{
is outlier
}
