% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{LOESS_norm}
\alias{LOESS_norm}
\title{LOESS normalization}
\usage{
LOESS_norm(
  object,
  feature_platform = "PLATFORM",
  QC_ID_pattern = "MTRX",
  span = 0.75,
  degree = 2,
  test = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Metabolite object. In the feature annotation slot `feature`, a platform column should be provided for metabolite measurement platform (eg. `PLATFORM`). The values in the `PLATFORM` column (eg. `Neg`, `Polar`, `Pos Early`, and `Pos Late`) are column names in the sample annotation `sample` to determine the batches of samples.}

\item{feature_platform}{The column name of feature platform for metabolite measurements (eg. `PLATFORM`).}

\item{QC_ID_pattern}{A character pattern to determine QC samples. Default value: "MTRX".}

\item{span}{default value 0.4}

\item{degree}{default value 2}

\item{test}{test the function for the first 20 columns.}

\item{verbose}{print log information.}
}
\description{
Normalization data by fitting locally estimated scatterplot smoothing (LOESS) on QC samples in each batch. For each metabolite, the values (eg. raw peak area data) were divided by the median value of QC samples in that batch. QC samples and metabolite batches should be specified (see parameters below).
}
\examples{
\dontrun{
d <- QCmatrix_norm(object = df)
}
}
\seealso{
\code{\link{batch_norm}}
}
