% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{maplet2Metabolite}
\alias{maplet2Metabolite}
\title{Convert maplet data to a Metabolite object}
\usage{
maplet2Metabolite(object, featureID = "name", sampleID = "sample")
}
\arguments{
\item{object}{Data in the maplet package, a SummarizedExperiment data object.}

\item{featureID}{feature ID column name in the maplet data (-> elementMetadata -> listData).}

\item{sampleID}{sample ID column name in the maplet data (-> colData -> listData).}
}
\value{
A Metabolite object.
}
\description{
Convert maplet data to a Metabolite object
}
