% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{merge_data}
\alias{merge_data}
\title{merge two Metabolite objects}
\usage{
merge_data(object_X = NULL, object_Y = NULL, all = TRUE, verbose = TRUE)
}
\arguments{
\item{object_X}{The first Metabolite object.}

\item{object_Y}{The second Metabolite object.}

\item{all}{logical; all = TRUE: keep all metabolites; all = FALSE, 
keep common metabolites that were present in both datasets.}

\item{verbose}{print log information.}
}
\value{
A Metabolite object after merging with slots: assayData, featureData, and sampleData.
}
\description{
Merge two Metabolite objects.
}
