% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Metabolite2maplet}
\alias{Metabolite2maplet}
\title{Convert a Metabolite object to maplet data}
\usage{
Metabolite2maplet(object)
}
\arguments{
\item{object}{A Metabolite object}
}
\value{
Data in the maplet package, a SummarizedExperiment data object.
}
\description{
Convert a Metabolite object to maplet data
}
