% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{outlier_rate}
\alias{outlier_rate}
\alias{outlier_rate.default}
\alias{outlier_rate.data.frame}
\alias{outlier_rate.Metabolite}
\title{outlier rate}
\usage{
outlier_rate(object, nSD)

\method{outlier_rate}{default}(object, nSD = 5)

\method{outlier_rate}{data.frame}(object, nSD = 5)

\method{outlier_rate}{Metabolite}(object, nSD = 5)
}
\arguments{
\item{object}{An object, vector, data.frame, data.table or Metabolite.}

\item{nSD}{N times of the SD as outliers.}
}
\value{
Returns a vector of the outlier rate.

A data.table of outlier rate.
}
\description{
Calculate outlier rate.
}
\examples{
# for a Metabolite object
data(df_plasma)
v <- outlier_rate(df_plasma)

}
