% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_Metabolite}
\alias{plot_Metabolite}
\title{plot a Metabolite object}
\usage{
plot_Metabolite(
  object,
  plot = "boxplot",
  x = "NEG",
  y = "value",
  color = "NEG",
  shape = "NEG",
  feature_name = NULL,
  random_select = 16,
  size = 0.6
)
}
\arguments{
\item{object}{A Metabolite object.}

\item{plot}{type of plot}

\item{x}{The x-axis coordinate.}

\item{y}{The y-axis coordinate.}

\item{color}{A column in `@sampleData` to show the color of points.}

\item{shape}{A column in `@sampleData` to show the shape of points.}

\item{feature_name}{A vector of selected metabolites to plot. If NULL, will randomly select 16 (default) metabolites to plot.}

\item{random_select}{An integer, number of randomly selected metabolites to plot.}

\item{size}{Point size.}
}
\description{
Plot a Metabolite object including boxplot (more to add.).
}
\examples{
\dontrun{
p <- plot_Metabolite(df_m_PCA, plot = "boxplot")
p
}

}
