% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot PCA}
\usage{
plot_PCA(object, color = "NEG", shape = "NEG", size = 1.5)
}
\arguments{
\item{object}{A Metabolite object.}

\item{color}{A column in `@sampleData` to show the color of points.}

\item{shape}{A column in `@sampleData` to show the shape of points.}

\item{size}{Point size.}
}
\value{
PCA plot.
}
\description{
Plot first two principal components.
}
