% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_QC}
\alias{plot_QC}
\title{quality control visualization}
\usage{
plot_QC(object, nSD = 5)
}
\arguments{
\item{object}{A Metabolite object.}

\item{nSD}{Define the N times of the SD as outliers.}
}
\value{
QC metrics
}
\description{
This function will plot QC results
}
