% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_ROC}
\alias{plot_ROC}
\title{ROC}
\usage{
plot_ROC(
  object = NULL,
  y = NULL,
  x = NULL,
  model_a = NULL,
  model_b = NULL,
  lab = NULL
)
}
\arguments{
\item{object}{A Metabolite object.}

\item{y}{A column name for the disease (0, 1)}

\item{x}{One variable name (if x is provided, model_a and model_b should be NULL or vice versa).}

\item{model_a}{Column names for model a (one or more covariates, as the first model).}

\item{model_b}{Column names for model b (one or more covariates, as the second model).}

\item{lab}{Title (eg. "BIOCHEMICAL"), default value is x.}
}
\value{
ROC.
}
\description{
Plot Receiver Operating Characteristic (ROC) curve for metabolites with or without covariates
}
