% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_tsne}
\alias{plot_tsne}
\title{plot tSNE}
\usage{
plot_tsne(object, color = "NEG", shape = "NEG", size = 1.5)
}
\arguments{
\item{object}{A Metabolite object.}

\item{color}{A column in `@sampleData` to show the color of points.}

\item{shape}{A column in `@sampleData` to show the shape of points.}

\item{size}{Point size.}
}
\value{
tSNE plot.
}
\description{
Plot t-distributed stochastic neighbor embedding. See more details in \code{\link[M3C]{tsne}}.
}
