% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_volcano}
\alias{plot_volcano}
\title{volcano plot for regression results}
\usage{
plot_volcano(
  fit,
  x = "estimate",
  y = "p.value",
  p.value_log10 = TRUE,
  color = "outcome",
  label = "term",
  highlight = "significant",
  x_lab = "Effect size",
  y_lab = "-log10(P value)"
)
}
\arguments{
\item{fit}{regression summary results.}

\item{x}{The x-axis column, eg. effect size.}

\item{y}{The y-axis column, eg. p value.}

\item{p.value_log10}{whether to transforme p.value by -log10.}

\item{color}{A column in fit to show different point colors. Set as NULL to turn off the color argument.}

\item{label}{A column in fit to label points.}

\item{highlight}{A column in fit to show the points to highlight. Values as 1 are hightlighted.}

\item{x_lab}{labels for x-axis.}

\item{y_lab}{labels for y-axis.}
}
\description{
volcano plot for regression results
}
\examples{
\dontrun{
p <- plot_volcano(fit_lm, color = NULL)
p
}

}
