% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{rescale}
\alias{rescale}
\title{rescale based on conversion factors}
\usage{
rescale(
  object,
  conversion_factor_data = NULL,
  QC_ID_pattern = "MTRX",
  verbose = TRUE
)
}
\arguments{
\item{object}{A Metabolite object. In the `featureData`, `conversion_factor_ID` column should be created to match with conversion_factor_data.}

\item{conversion_factor_data}{A data set with columns `conversion_factor_ID` and `conversion_factor_value`.}

\item{QC_ID_pattern}{A character pattern to determine QC samples. Default value: "MTRX". Skip QC samples when rescale (median value is already 1).}

\item{verbose}{print log information.}
}
\description{
Rescale metabolite data based on a conversion factor file
}
\examples{

\dontrun{
d <- rescale(object)
}

}
