% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{row_missing_rate}
\alias{row_missing_rate}
\alias{row_missing_rate.default}
\alias{row_missing_rate.Metabolite}
\title{row missing rate}
\usage{
row_missing_rate(object)

\method{row_missing_rate}{default}(object)

\method{row_missing_rate}{Metabolite}(object)
}
\arguments{
\item{object}{An object, data.frame, data.table or Metabolite.}
}
\value{
Returns a vector of the missing rate for each row

A data.table of row missing rate.
}
\description{
Calculate row missing rate -- sample missingness.
}
\examples{
# for a Metabolite object
data(df_plasma)
v <- row_missing_rate(df_plasma)

}
