% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{sampleData<-}
\alias{sampleData<-}
\alias{sampleData<-,Metabolite-method}
\title{set sampleData}
\usage{
sampleData(object) <- value

\S4method{sampleData}{Metabolite}(object) <- value
}
\arguments{
\item{object}{A Metabolite object.}

\item{value}{The new sampleData.}
}
\value{
A data.table of sampleData.
}
\description{
Accessors for Metabolite object. `sampleData<-` will update the sampleData in the Metabolite object.
}
