% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_data_xls}
\alias{save_data_xls}
\title{Save metabolite data in an Excel file}
\usage{
save_data_xls(object, file = "")
}
\arguments{
\item{object}{A Metabolite object}

\item{file}{Output file to save the metabolite measurements 
(suffixes: ".xlsx", "_logs.txt).}
}
\value{
No return value.
}
\description{
Save metabolite data in an Excel file
}
