% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{show,Metabolite-method}
\alias{show,Metabolite-method}
\title{Print a Metabolite class object}
\usage{
\S4method{show}{Metabolite}(object)
}
\arguments{
\item{object}{A Metabolite object.}
}
\value{
print a Metabolite object.
}
\description{
Print a Metabolite class object
}
