% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/process.R
\name{column_missing_rate}
\alias{column_missing_rate}
\alias{column_missing_rate.default}
\alias{column_missing_rate.Metabolite}
\title{column missing rate}
\usage{
column_missing_rate(object)

\method{column_missing_rate}{default}(object)

\method{column_missing_rate}{Metabolite}(object)
}
\arguments{
\item{object}{An object, data.frame, data.table or Metabolite.}
}
\value{
Returns a vector of the missing rate for each column
}
\description{
Calculate column missing rate -- metabolite missingness.
}
\examples{
\dontrun{
# for a data.frame or data.table
v <- column_missing_rate(object = df)

# to skip the first column (eg. ID)
v <- column_missing_rate(object = df[, -1])
}

\dontrun{

# for a Metabolite object
v <- column_missing_rate(object)

}

}
