% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{pareto_scale}
\alias{pareto_scale}
\title{pareto scale transformation}
\usage{
pareto_scale(x)
}
\arguments{
\item{x}{A vector}
}
\description{
pareto scale transformation
}
\examples{

\dontrun{
v <- paretoscale(x)
}

}
