% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{save_data}
\alias{save_data}
\title{Save metabolite data}
\usage{
save_data(object, file = "")
}
\arguments{
\item{object}{A Metabolite object}

\item{file}{Output file to save the metabolite measurements 
(suffixes: "_assay.csv", "_feature_annotation.csv", "_sample_annotation.csv",).}
}
\description{
Save metabolite data in three separate csv files
}
